Fix error during cypress launch
